Create Sounds namespace with MacOS and Mute classes
