+++ test/retest/sounds/sounds_interface.rb
@@ -1,0 +1,13 @@
+require_relative '../runners/observable_runner'
+
+module Retest
+  module Sounds
+    module SoundsInterfaceTests
+      include Runners::ObserverInterfaceTests
+
+      def test_interface
+        assert_respond_to @subject, :play
+      end
+    end
+  end
+end
