+++ test/retest/sounds_test.rb
@@ -1,36 +1,10 @@
 require 'test_helper'
-require_relative 'runners/observable_runner'
 
 module Retest
   class SoundsTest < MiniTest::Test
-    include Runners::ObserverInterfaceTests
-
-    def setup
-      @subject = Sounds.new
-    end
-
-    class FakeThread
-      def initialize(&block)
-        block.call
-      end
-    end
-
-    def test_play_tests_pass
-      kernel = MiniTest::Mock.new
-      kernel.expect(:system, true, ['afplay', '/System/Library/Sounds/Funk.aiff'])
-
-      Sounds.new(kernel: kernel, thread: FakeThread).play(:tests_pass)
-
-      kernel.verify
-    end
-
-    def test_play_tests_fail
-      kernel = MiniTest::Mock.new
-      kernel.expect(:system, true, ['afplay', '/System/Library/Sounds/Sosumi.aiff'])
-
-      Sounds.new(kernel: kernel, thread: FakeThread).play(:tests_fail)
-
-      kernel.verify
+    def test_sounds_for
+      assert_instance_of Sounds::MacOS, Sounds.for(Options.new(['--notify']))
+      assert_instance_of Sounds::Mute,  Sounds.for(Options.new([]))
     end
   end
 end
